redesigning project structure; removing old code
