--- constants.py
@@ -1,5 +1,0 @@
-CORPUS_PATH = 'data/enwiki-20171001-pages-meta-current-withlinks-processed/AE/wiki_00.bz2'
-
-RETRIEVER_VECTOR_REPRESENTATION_PATH = ''
-REASONER_MODEL_PATH = 'models/albert-binary-classifier'
-READER_MODEL_PATH = 'models/squad-qa-model'
