--- docker-compose.yml
@@ -1,18 +1,0 @@
-version: '3'
-services:
-    graph-qa:
-        image: graph-qa
-        container_name: graph-qa
-        build:
-            context: .
-            dockerfile: graph-qa.Dockerfile
-        volumes:
-            - .:/graph-qa
-        tty: True
-        ports:
-            - "127.0.0.1:8888:8888"
-            #- "127.0.0.1:6379:6379"
-            
-    redisdb:
-        image: redis:5.0.5
-        container_name: redis-db
