+++ docker/docker-compose.yml
@@ -1,0 +1,2 @@
+version: '3'
+services:
