+++ graphqa/core/pipeline.py
@@ -1,0 +1,6 @@
+class Pipeline:
+    def __init__(self):
+        pass
+
+    def __call__(self, *args, **kwargs):
+        pass
