+++ graphqa/core/reader/__init__.py
@@ -1,0 +1,1 @@
+from .abstract_reader import AbstractReader
