+++ graphqa/core/reasoner/__init__.py
@@ -1,0 +1,1 @@
+from .abstract_reasoner import AbstractReasoner
