+++ graphqa/core/reasoner/abstract_reasoner.py
@@ -1,0 +1,12 @@
+
+from abc import ABC, abstractmethod
+
+
+class AbstractReasoner(ABC):
+    @abstractmethod
+    def load(self, path):
+        raise NotImplementedError()
+
+    @abstractmethod
+    def rank(self, question, paragraphs):
+        raise NotImplementedError()
