+++ graphqa/core/retriever/__init__.py
@@ -1,0 +1,1 @@
+from .abstract_retriever import AbstractRetriever
