--- modules/reader/abstract_reader.py
@@ -1,8 +1,0 @@
-
-class AbstractReader:
-
-    def __init__(self):
-        pass
-
-    def extract_answer(self, question, paragraphs):
-        pass
