--- modules/reasoner/abstract_reasoner.py
@@ -1,8 +1,0 @@
-
-class AbstractReasoner:
-
-    def __init__(self):
-        pass
-    
-    def rank(self, question, paragraphs):
-        pass
