--- modules/retriever/abstract_retriever.py
@@ -1,9 +1,0 @@
-class AbstractRetriever:
-    def __init__(self):
-        pass
-
-    def load_corpus(self, corpus_path):
-        pass
-
-    def retrieve(self, question, paragraphs):
-        pass
