--- requirements.txt
@@ -1,10 +1,0 @@
-jupyter
-transformers
-pandas
-numpy
-ordered-set
-tqdm
-sklearn
-simpletransformers
-nltk
-redis
