+++ setup.py
@@ -1,0 +1,9 @@
+from setuptools import setup, find_packages
+
+setup(
+    name='graph-qa',
+    version='0.1',
+    packages=find_packages(),
+    include_package_data=True,
+    zip_safe=False
+)
